Restructure msgbus
