+++ nautilus_core/infrastructure/Cargo.toml
@@ -21,6 +21,7 @@ log = { workspace = true }  # Required until Cython gone
 rmp-serde = { workspace = true }
 rust_decimal = { workspace = true }
 semver = { workspace = true }
+serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
 tracing = {workspace = true }
